Breaking out storage model classes into separate files
